add sweetalert for deletion confirmation

+++ resources/views/users/index.blade.php
@@ -47,7 +47,7 @@
                                                 </a>
                                                 <div class="dropdown-menu dropdown-menu-right dropdown-menu-arrow">
                                                     <a class="dropdown-item" href="/user/edit/{{ $user->id }}">Edit</a>
-                                                    <a class="dropdown-item" href="/user/delete/{{ $user->id }}">Delete</a>
+                                                    <a class="dropdown-item warning-confirm" href="/user/delete/{{ $user->id }}">Delete</a>
                                                 </div>
                                             </div>
                                         </td>
